Step 6: Lift flush and close, with readRevisionNumber as abstract
Moves the flush and close implementations to AbstractFormatDriver, using the
already-moved flushLock and downstream. The readRevisionNumber method is now
abstract, implemented differently by each subclass.
