+++ bosk-mongo/src/main/java/works/bosk/drivers/mongo/internal/PandoFormatDriver.java
@@ -148,19 +148,6 @@ final class PandoFormatDriver<R extends StateTreeNode> extends AbstractFormatDri
 	}
 
 	@Override
-	public void flush() throws IOException, InterruptedException {
-		flushLock.awaitRevision(readRevisionNumber());
-		LOGGER.debug("| Flush downstream");
-		downstream.flush();
-	}
-
-	@Override
-	public void close() {
-		LOGGER.debug("+ close()");
-		flushLock.close();
-	}
-
-	@Override
 	BsonStateAndMetadata loadBsonStateAndMetadata() throws UninitializedCollectionException {
 		List<BsonDocument> allParts = new ArrayList<>();
 		try (MongoCursor<BsonDocument> cursor = collection
@@ -639,7 +626,7 @@ final class PandoFormatDriver<R extends StateTreeNode> extends AbstractFormatDri
 	 * @return Non-null revision number as per the database.
 	 * If the database contains no revision number, returns {@link Formatter#REVISION_ZERO}.
 	 */
-	private BsonInt64 readRevisionNumber() throws FlushFailureException {
+	protected BsonInt64 readRevisionNumber() throws FlushFailureException {
 		LOGGER.debug("readRevisionNumber");
 		try {
 			try (MongoCursor<BsonDocument> cursor = collection
